chore: Configure DWO with tlsCertificateConfigmapRef when certificates imported to che-operator
