+++ bundle/next/eclipse-che/manifests/che-operator.clusterserviceversion.yaml
@@ -86,7 +86,7 @@ metadata:
     categories: Developer Tools
     certified: "false"
     containerImage: quay.io/eclipse/che-operator:next
-    createdAt: "2026-06-16T14:03:55Z"
+    createdAt: "2026-06-23T10:08:27Z"
     description: A Kube-native development solution that delivers portable and collaborative
       developer workspaces.
     features.operators.openshift.io/cnf: "false"
@@ -108,7 +108,7 @@ metadata:
     operatorframework.io/arch.amd64: supported
     operatorframework.io/arch.arm64: supported
     operatorframework.io/os.linux: supported
-  name: eclipse-che.v7.119.0-990.next
+  name: eclipse-che.v7.120.0-991.next
   namespace: placeholder
 spec:
   apiservicedefinitions: {}
@@ -1153,7 +1153,7 @@ spec:
       name: gateway-authorization-sidecar
     - image: quay.io/che-incubator/header-rewrite-proxy:latest
       name: gateway-header-sidecar
-  version: 7.119.0-990.next
+  version: 7.120.0-991.next
   webhookdefinitions:
     - admissionReviewVersions:
         - v1
