+++ pkg/deploy/server/chehost_reconciler_test.go
@@ -55,8 +55,7 @@ func TestSyncService(t *testing.T) {
 	assert.Nil(t, err)
 
 	service := &corev1.Service{}
-	done, err = deploy.Get(ctx, types.NamespacedName{Name: deploy.CheServiceName, Namespace: "eclipse-che"}, service)
-	assert.True(t, done)
+	err = ctx.ClusterAPI.Client.Get(context.TODO(), types.NamespacedName{Name: deploy.CheServiceName, Namespace: "eclipse-che"}, service)
 	assert.Nil(t, err)
 
 	assert.Equal(t, service.Spec.Ports[0].Name, "http")
